Small changes in submissions models

+++ Dto/Frontend/TelegramDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int StudentId { get; set; }
-        public int TelegramId { get; set; }
+        public ulong TelegramId { get; set; }
         public string Language { get; set; } = null!;
         public bool ScheduleCompact { get; set; }
         public DateOnly? LastCongratulations { get; set; }
